Refactor everything from model.py to more sensible locations
